change the value types

+++ crates/api_models/src/routing.rs
@@ -705,14 +705,13 @@ pub struct ToggleDynamicRoutingPath {
 #[derive(serde::Serialize, serde::Deserialize, Debug, Clone, ToSchema)]
 pub struct EliminationRoutingConfig {
     pub params: Option<Vec<DynamicRoutingConfigParams>>,
-    // pub labels: Option<Vec<String>>,
     pub elimination_analyser_config: Option<EliminationAnalyserConfig>,
 }
 
 #[derive(serde::Serialize, serde::Deserialize, Debug, Clone, ToSchema)]
 pub struct EliminationAnalyserConfig {
-    pub bucket_size: Option<u32>,
-    pub bucket_ttl_in_mins: Option<f64>,
+    pub bucket_size: Option<u64>,
+    pub bucket_ttl_in_mins: Option<u64>,
 }
 
 impl Default for EliminationRoutingConfig {
@@ -721,7 +720,7 @@ impl Default for EliminationRoutingConfig {
             params: Some(vec![DynamicRoutingConfigParams::PaymentMethod]),
             elimination_analyser_config: Some(EliminationAnalyserConfig {
                 bucket_size: Some(5),
-                bucket_ttl_in_mins: Some(2.0),
+                bucket_ttl_in_mins: Some(2),
             }),
         }
     }
